exec directive should not open a win console ever
Previously the exec directive would open a windows console
when run from the windowed variant of the dgenerate
executable on Windows.

A creation flag needs to be supplied to popen to prevent this
from ever happening when running on windows.

Additionally --no-stdin should be allowed with --file as it
is harmless and allows for: \\exec dgenerate --file myfile.dgen

To function properly in a config file without too much effort