webcache for download directive alt directorys
Use the webcache as a filepointer store to record
where files were downloaded to when they were
downloaded to an explicit location with the download
directive or template function.

This removes the need for a request to be made every
time to figure out the filename of a download before
checking if it exists or not.

At least until the filepointer expires in the webcache.