Prompt weight before pipeline on device
Prompt weighters should not expect text encoder
to already be on the device, and should move it
to the desired device before using it.

This saves memory while embed generation is
happening before the pipeline is actually called.

The other modules do not need to be on the device
in order to generate embeds, and if they do,
then the embed generator should move them there.

This also fixes a global state issue with llm4gens
pre embed generation memory guarding, where it
may have tried to remove any cached pipeline
on the device back to the cpu to make space.

llm4gen now always moves the text_encoder before generating
embeds, so a device mismatch error cannot happen in low
memory condition.