Implement llm4gen clip weighting, rankgen prompt specification #12
Allow clip embeddings to be weighted using sd-embed or compel

Allow T5 rankgen encoder prompt to be explicitly specified with
dgenerate --second-prompts

If weighting is occuring and --second-prompts is not used to
specify the rankgen prompt, use the primary prompt with
all weighting syntax removed.

This uses a new prompt weighter plugin API (overridable method)
for specifying extended pipeline prompt arguments, specifically
when they are not supported by the model that the embed
generation is occuring for by default.

This allows you to indicate that prompt_2, negative_prompt_2,
prompt_3, or negative_prompt_3 are supported.

dgenerate can then generate these pipeline arguments via
--second-prompts, --third-prompts, or --second-model-second-prompts.

Normally PipelineDiffusionWrapper will reject generating
invalid arguments for a diffusion pipeline, it now integrates with this
API to allow you to specify that your prompt weighter can handle
the extra prompt inputs when the pipeline signature itself
does not support it.