Work on wheel packaging fixes for quant #14
torchao and bitsandbytes are now extras.

The extra "quant" installs all compatible quantization backends
on your platform.

If any of these extras are missing from the built wheel, it is
infered that a default install method is not supported on your platform

Document the possiblity of bitsandbytes pre-release multiplatform
wheels being an option, and how to install for ROCm and MacOS using
pipx or standard virtual env.