setup.py fixes for win-installer, installer build fixes
pyinstaller needs to include static data files related to dealing
with language codes and language specification
for the translate prompt upscaler plugin #24

Need to include bitsandbytes package in the
frozen environment #14

cleanup setup.py wheel build a little more.

Idea is that if the package extras do not exist in the
wheel for your platform, then they are currently
unsupported for your platform.