Automatic1111 metadata tool, --sub-command auto1111-metadata #39
This sub-command converts a dgenerate config stored on an image with
--output-metadata into Automatic1111 format, and stores it back on
to the image or on to a copy of that image.

It can also read the output of --output-configs, (.dgen config files)
to create this metadata.

This is useful for preparing images for upload to CivitAI

Model hashing is currently up front, and works only when
the image was generated from single checkpoint files, and
for sub-models that were specified as single checkpoint files.

Hugging face slugs are passed as is into the metadata model names,
and a hash is not provided.

If your generation links to models such as CivitAI links, they
will be looked for in dgenerates web cache or downloaded, and
then hashed.

A hash can be provided for: Primary Model, VAE, LoRAs, ControlNets, and
TextualInversions (embeddings), when these sub-models are used as a
single file on disk, or a link to a model such as a CivitAI link.