--sigmas/--sdxl-refiner-sigmas expression syntax #33
These arguments now understand values beginning with "expr: "
to represent a python expression involving "sigmas" generated
by the scheduler

"sigmas" is represented as a numpy array, which provides
an easy way to scale or manipulate sigmas produced by the
models scheduler

for instance --sigmas "expr: sigmas * 0.95"

In this expression environment, numpy is available in
the familiar namespace "np", meaning that functions
such as "np.linspace" can be used in sigmas expressions
without the use of dgenerates config shell, i.e. when
you are using the CLI tool from the system shell

Python builtins like "min", "max" etc. are also
available.

This functionality is provided by the expression parsing
package asteval which is used elsewhere for dgenerates shell

CSV and expression can be mixed together when specifying
sigmas in the combinatorial sense

--sigmas "expr: sigmas * 0.95" "1.0,0.9,0.8"

Added a documentation section describing this feature,
and examples for Flux, SD3, and SDXL