scheduler continuity fix, reset scheduler to default when needed
The scheduler URI is not part of the pipelines cache key as it
can be modified on the fly.

The problem is, if scheduler URI is None after we have set the
scheduler on the pipeline previously (indicating to use the default), the
pipeline would continue to use the last set scheduler
as it is still alive in the cache with that scheduler set

we need to store the intial scheduler the pipeline was loaded with
somewhere on the pipeline using an attribute tag, so that when we specify no
scheduler URI (i.e. use the default), the default scheduler for the
pipeline can be reset back onto the pipeline