(amend missing changes) latents processors similar to image processors #41
Add plugin framework for latents processors, which can be used to manipulate
incoming or outgoing latents

implement --latents-processors, --img2img-latents-processors, and --latents-post-processors

--latents-processors affects raw latents --image-seeds "latents: ..." and --image-seeds "img2img.png;latents ..."

--img2img-latents-processors affects img2img latents --image-seeds "img2img_latents.py" (decoded by VAE)

--latents-post-processors can post process latents output when --image-format (pt, pth, or safetensors)

--seed-image-processors can now act on incoming latents passed as img2img data after they
are decoded by the receiving VAE

Resizing operations are now applied to incoming latents passed as img2img data after they
are decoded by the receiving VAE

DiffusionPipelineWrapper now implements internal resizing for image inputs based on width/height
provided to diffusion arguments for models where input image size is related to output size,
previously this was ignored and the output size was just determined by the input size.  Usually
images are resized before they ever get to the pipeline wrapper so this was not a problem,
this is just a nicety for library users that want to use this class. Stable Cascade which uses an
image encoder to create prompt like embeddings does not have its inputs undergo this
resizing process as the input images are used as a content reference and not a
noise starting point.

Integrate new options with recipe form Console UI on supported models

Add to Console UI: Insert URI -> Latents Processor URI

Add to Console UI: Insert Image Seed now has fields for raw latents and simple IP Adapter image specifications.

Implement "interposer" processor, and "noise", and "scale"

Better input validation for latents channels and decode dimensions, avoid rotary embedding
errors when user specifies an output size that does not match the dimensions of the latents

Flux latents are now stored on disk / output in unpacked format, pipeline wrapper now expects all latents
input in unpacked format for consistency, this also allows for the image dimension to be encoded in the
dimension of the latents for easier validation.  It is easier to pack the Flux latents for pipeline
input than to unpack them for processing than re-pack them.

Adetailer pipeline workflow can now accept img2img latents, as they will be decoded into an image before being processed.

Adetailer cannot output to latents as it requires post processing in pixel space.

Add documentation and examples

Various fixes and refactoring along the way