adetailer "class-filter" / --adetailer-class-filter, --adetailer-model-masks, various adetailer improvements
Add "class-filter" argument to adetailer image processor and --adetailer-detectors URI

This allows you to specify which object classes to include during detection, by ID (integer) or name (string).

Add "model-masks" argument to adetailer image processor / --adetailer-detectors URI, as well as
--adetailer-model-masks argument.

This allows you to specify that masks generated by the YOLO model should be used
if available.

This bypasses mask generation from the detection bounding box and related settings
in favor of the mask generated by the YOLO model, which some models are capable of generating.
By default this value is False, meaning that the bounding box will be used to generate the mask
unless this setting is turned on.  If the model does not actually generate a mask, then mask
generation will fall back to the bounding box method.

Implement image processor "yolo" for previewing YOLO model detections on an image, as well as
generating masks. This image processor can generate masks identical to those used by adetailer
internally, minus the gaussian blur and dilation processing on the mask.  This processor implements
many of the options that adetailer does, such as the class filter / index filter and model masks options,
and can be used to preview what adetailer is going to affect, or to generate inpainting masks manually.

Implement the image processor "gaussian-blur" for applying a gaussian blur to an image using cv2.

Implement the image processor "dilate" for applying a dilation to an image using cv2.

Fix issues loading adetailer models from URLs.

Various fixes to the plugin documentation system and internal text processing.

Doc string cleanup.

Write more test coverage for the various URIs that dgenerate can parse.