global support for HF Blob links for single files, global download pbars
Everywhere a single file can be loaded as a model / config,
Hugging Face blob links are now supported.

This includes image processors etc. where a model can be specified,
basically anywhere that would otherwise be able to accept a path to
a singular file on disk, or a direct download URL to a file.

Comprehensive --offline-mode refactor for complete firewalling and
cache only operation, this now affects any download activity at all
and not just the hugging face cache

Library API supports very granular control over offline mode,
as well as total global control

Hack huggingface_hub progress bars to aggregate multithreaded
multithreaded progress bars into a single progress bar,
this is way easier for every terminal to support and provides
the total downloaded amount as a sort of spinner to
indicate something is at least happening

Hack huggingface_hub so that every download gets a progress bar,
period, it seems to have trouble with this