allow user to disable scheduler config cloning when using --scheduler/--second-model-scheduler
For --scheduler/--second-model-scheduler, you can now use the argument
"clone-config" with every scheduler.

This argument defaults to True, and indicates whether to clone
the models original scheduler config into the alternate scheduler
or not.

If you set this to False, the defaults displayed by
--scheduler/--second-model-scheduler helpargs will be
used in entirety.

This allows complete manual configuration of the scheduler
without any interference from the models original scheduler
configuration if desired.

Add more information to scheduler documentation section to
cover this usage.