--inpaint-crop, --inpaint-crop-paddings, --inpaint-crop-feathers, --inpaint-crop-masked #32
Pipeline built in's for autocropping to mask and processing the region at a higher resolution

Add examples which demonstrate identical results to implementation with image processors

This is more compatible with processing animated inputs with automatic feature detection,
just a little less configurable.

Using --inpaint-crop alone uses the default value of --inpaint-crop-paddings 50 (50 pixels),
no crop masking.

using other --inpaint-crop* arguments implies --inpaint-crop, so they can be used alone
and still trigger this mode.

--inpaint-crop-masked and --inpaint-crop-feathers are mutually exclusive, as feathering
generates a crop mask instead of using your original mask.

--inpaint-crop-paddings and --inpaint-crop-feathers are combinatoric and accept multiple values
if desired for variations.

feather refers to alpha feathering the edges of the pasted image back onto the background,
this defaults to 0 (off) but may be desired in some cases.

padding refers to the amount of padding around the detected mask bounds (white area sum total),
this determines how much padding is added during the crop to mask bounding box.

--inpaint-crop-masked indicates that the mask should be applied on paste, effectively removing
any non-generated content from the content that is pasted back over the image. You may wish to
use this argument if the upscale of the cropped region is so significant that it introduces
small artifacts into the non-generated background area when it is scaled back down to fit
over the original image.