Fixes for SD3/Flux output dimension in img2img/controlnet mode
We should not use default width/height as the output dimension for
SD3/Flux when an image input is used, we should use the size of
the input image (refactor bug)

We need to setup the same "upscaling/downscaling" logic for SD3/Flux controlnet
pipelines without additional img2img inputs. i.e. pipeline width/height
params do not need to be the same size as the image input, and we do
not need to pre-resize the input images.

We need to not resize input images before they get to the
DiffusionPipelineWrapper with SD3/Flux in the case of controlnets,
missed this on previous refactor. (RenderLoop)

For other pipelines, we should resize them, image processors usually
run after resize, which is good for providing proper depth/canny
controlnet images etc. (in default configuration) to pipelines
where input size is related to output size.

Since these images will not be pre-resized for SD3/Flux before
entering the pipeline, this pre-resize / post-resize logic is
not important for preserving control image or img2img content.

Also simplify example runner and provide checkpoint system
for making local hand testing easier.